fix: page.tsx 서버 컴포넌트로 HomeClientWrapper로 분리
